+++ web/tests/stream_data.test.cjs
@@ -23,6 +23,8 @@ const stream_settings_data = zrequire("stream_settings_data");
 const user_groups = zrequire("user_groups");
 const {initialize_user_settings} = zrequire("user_settings");
 
+const bot_data = mock_esm("../src/bot_data");
+
 const current_user = {};
 set_current_user(current_user);
 const realm = {};
@@ -101,6 +103,7 @@ function test(label, f) {
         people.init();
         people.add_active_user(me);
         people.initialize_current_user(me.user_id);
+        current_user.user_id = me.user_id;
         stream_data.clear_subscriptions();
         user_groups.initialize({
             realm_user_groups: [
@@ -1167,23 +1170,23 @@ test("can_post_messages_in_stream", ({override}) => {
         history_public_to_subscribers: false,
         stream_post_policy: settings_config.stream_post_policy_values.admins.code,
     };
-    override(current_user, "is_admin", false);
+    me.is_admin = false;
     assert.equal(stream_data.can_post_messages_in_stream(social), false);
 
-    override(current_user, "is_admin", true);
+    me.is_admin = true;
     assert.equal(stream_data.can_post_messages_in_stream(social), true);
 
     social.stream_post_policy = settings_config.stream_post_policy_values.moderators.code;
-    override(current_user, "is_moderator", false);
-    override(current_user, "is_admin", false);
+    me.is_moderator = false;
+    me.is_admin = false;
 
     assert.equal(stream_data.can_post_messages_in_stream(social), false);
 
-    override(current_user, "is_moderator", true);
+    me.is_moderator = true;
     assert.equal(stream_data.can_post_messages_in_stream(social), true);
 
     social.stream_post_policy = settings_config.stream_post_policy_values.non_new_members.code;
-    override(current_user, "is_moderator", false);
+    me.is_moderator = false;
     me.date_joined = new Date(Date.now());
     override(realm, "realm_waiting_period_threshold", 10);
     assert.equal(stream_data.can_post_messages_in_stream(social), false);
@@ -1191,7 +1194,7 @@ test("can_post_messages_in_stream", ({override}) => {
     me.date_joined = new Date(Date.now() - 20 * 86400000);
     assert.equal(stream_data.can_post_messages_in_stream(social), true);
 
-    override(current_user, "is_guest", true);
+    me.is_guest = true;
     assert.equal(stream_data.can_post_messages_in_stream(social), false);
 
     social.stream_post_policy = settings_config.stream_post_policy_values.everyone.code;
@@ -1367,34 +1370,45 @@ test("can_access_stream_email", ({override}) => {
         is_muted: false,
         invite_only: true,
         history_public_to_subscribers: false,
+        stream_post_policy: settings_config.stream_post_policy_values.admins.code,
     };
-    override(current_user, "is_admin", false);
-    assert.equal(stream_data.can_access_stream_email(social), true);
-
-    override(current_user, "is_admin", true);
-    assert.equal(stream_data.can_access_stream_email(social), true);
-
-    social.subscribed = false;
-    assert.equal(stream_data.can_access_stream_email(social), false);
-
-    social.invite_only = false;
-    assert.equal(stream_data.can_access_stream_email(social), true);
-
-    override(current_user, "is_admin", false);
-    assert.equal(stream_data.can_access_stream_email(social), true);
-
-    override(current_user, "is_guest", true);
-    assert.equal(stream_data.can_access_stream_email(social), false);
-
-    social.subscribed = true;
-    assert.equal(stream_data.can_access_stream_email(social), true);
-
-    social.is_web_public = true;
-    assert.equal(stream_data.can_access_stream_email(social), true);
-
-    social.subscribed = false;
-    assert.equal(stream_data.can_access_stream_email(social), true);
+    const bot = {
+        is_active: true,
+        user_id: 999,
+    };
+    const bot_user = {
+        email: "bot@zulip.com",
+        full_name: "Bot User",
+        user_id: 999,
+        is_bot: true,
+        role: settings_config.user_role_values.moderator.code,
+    };
+    people.add_active_user(bot_user);
+    override(bot_data, "get_all_bots_for_current_user", () => [bot]);
 
     page_params.is_spectator = true;
     assert.equal(stream_data.can_access_stream_email(social), false);
+    page_params.is_spectator = false;
+
+    social.subscribed = false;
+    assert.equal(stream_data.can_access_stream_email(social), false);
+    social.subscribed = true;
+
+    // Neither the user nor any of the bots they control can post message.
+    me.is_moderator = true;
+    assert.equal(stream_data.can_access_stream_email(social), false);
+
+    // user can but none of the bots they control can post message.
+    me.is_admin = true;
+    assert.equal(stream_data.can_access_stream_email(social), true);
+
+    // user can't but one of the bots they control can post message.
+    me.is_admin = false;
+    me.is_moderator = false;
+    social.stream_post_policy = settings_config.stream_post_policy_values.moderators.code;
+    assert.equal(stream_data.can_access_stream_email(social), true);
+
+    // Both the user and one of the bots they control can post message.
+    me.is_moderator = true;
+    assert.equal(stream_data.can_access_stream_email(social), true);
 });